Combine source handlers
