+++ tests/Functional/Client/AbstractClientTestCase.php
@@ -19,8 +19,7 @@ use SmartAssert\SourcesClient\Client;
 use SmartAssert\SourcesClient\ExceptionFactory;
 use SmartAssert\SourcesClient\RequestFactory;
 use SmartAssert\SourcesClient\RequestHandler\FileHandler;
-use SmartAssert\SourcesClient\RequestHandler\SourceAccessHandler;
-use SmartAssert\SourcesClient\RequestHandler\SourceMutationHandler;
+use SmartAssert\SourcesClient\RequestHandler\SourceHandler;
 use SmartAssert\SourcesClient\SourceFactory;
 use SmartAssert\SourcesClient\Tests\Functional\DataProvider\CommonNonSuccessResponseDataProviderTrait;
 use SmartAssert\SourcesClient\Tests\Functional\DataProvider\InvalidJsonResponseExceptionDataProviderTrait;
@@ -58,8 +57,7 @@ abstract class AbstractClientTestCase extends TestCase
 
         $this->client = new Client(
             new FileHandler($requestFactory, $serviceClient, $exceptionFactory),
-            new SourceMutationHandler($requestFactory, $serviceClient, $sourceFactory, $exceptionFactory),
-            new SourceAccessHandler($requestFactory, $serviceClient, $sourceFactory, $exceptionFactory),
+            new SourceHandler($requestFactory, $serviceClient, $sourceFactory, $exceptionFactory),
         );
     }
